requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.8.2
cryptography==45.0.3
easyocr==1.7.2
et-xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.5.1
imageio==2.37.0
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.1
lazy-loader==0.4
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer-six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.6.0
pyclipper==1.3.0.post6
pycparser==2.22
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.3.5
pytest-cov==6.1.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.26
torch==2.7.0+cu128
torchvision==0.22.0+cu128
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
